birdclef-2021/requirements.txt
pysoundfile
numpy
scipy
matplotlib
bokeh
pandas
